# Optimize query_cache_hit to reduce code size of the query hot path.