@Zoxc 2023-02-01
It makes sense that perf regresses since it runs with the profiler enabled. I've optimized `query_cache_hit` instead to avoid the extra function call and branch in the query system when profiling.