@Zoxc 2023-02-06
It looks like this is a bigger win after https://github.com/rust-lang/rust/pull/107667:
<table><tr><td rowspan="2">Benchmark</td><td colspan="1"><b>Before</b></th><td colspan="2"><b>After</b></th></tr><tr><td align="right">Time</td><td align="right">Time</td><td align="right">%</th></tr><tr><td>🟣 <b>clap</b>:check</td><td align="right">1.8384s</td><td align="right">1.8200s</td><td align="right">💚  -1.00%</td></tr><tr><td>🟣 <b>hyper</b>:check</td><td align="right">0.2624s</td><td align="right">0.2604s</td><td align="right"> -0.74%</td></tr><tr><td>🟣 <b>regex</b>:check</td><td align="right">1.0245s</td><td align="right">1.0113s</td><td align="right">💚  -1.28%</td></tr><tr><td>🟣 <b>syn</b>:check</td><td align="right">1.6461s</td><td align="right">1.6298s</td><td align="right"> -0.99%</td></tr><tr><td>🟣 <b>syntex_syntax</b>:check</td><td align="right">6.3677s</td><td align="right">6.3036s</td><td align="right">💚  -1.01%</td></tr><tr><td>Total</td><td align="right">11.1390s</td><td align="right">11.0252s</td><td align="right">💚  -1.02%</td></tr><tr><td>Summary</td><td align="right">1.0000s</td><td align="right">0.9900s</td><td align="right">💚  -1.00%</td></tr></table>