@rylev 2023-02-14
Calling this triaged as the regression is small

@rustbot label: perf-regression-triaged